dev fuckups

+++ models/DFHModel.h
@@ -69,8 +69,8 @@ public:
 private:
 	MPI_Comm comm;
     
-	int dist_mem_size;
-	int neighborSize;
+	size_t dist_mem_size;
+	size_t neighborSize;
 	// filenames
     char LocalRankString[8];
     char LocalRankFilename[40];
